RBXPRR-20: Update __all.ts
Add a notice for RBXPRR-27 (Support websockets) at the top of the file.

Add a method _isOriginAllowed that iterates a CORs rule's allowedOrigins
and tests each regex against the request origin.

Rewrite _applyCorsHeaders to now use corsWriter, it will do the
following:
  1. Fetch the CORs rule from corsWriter passing the context's `Request`
     object as an argument (used to determine request method, path etc.)
  2. If the rule does not exist we will return a true value, the return
     of this method determines if further down the proxy pipeline if the
     response from the proxy can overwrite these headers (except
     Access-Control-Allow-Origin), this will also return true if the
     CORs rule's allowResponseHeadersOverwrite is true.
  3. We will call on _isOriginAllowed, passing in the CORs rule's
     allowedOrigins and casting them to a RegExp[] (as it's a (string |
     RegExp)[]) and the request origin we captured from an outside
     context. If the result of this method is false, we will then check
     the CORs rule's allowRequestOriginIfNoAllowedOrigins field and if
     it's true we will proceed to Step 4, otherwise we will check if the
     setting corsApplyHeadersRegardlessOfOrigin is enabled, or we will
     just go back to Step 2.
  4. We can now apply the origin header from the client, this may be
     volatile if corsApplyHeadersRegardlessOfOriginHeader is enabled,
     and the request didn't actually have an Origin header, as this
     header in the response will just be `undefined` as a string. We
     then check if the allowedHeaders, allowedMethods, exposedHeaders,
     maxAge and allowCredentials are set on the CORs rule, and then
     apply the corresponding headers on the response. We will also set
     the `Vary` header to `Origin`, regardless of if the origin is
     actually present.
  5. Return the CORs rule's allowResponseHeadersOverwrite field.

Move the origin header check from the top of the invoke method (around
where the publicIp is initialized to directly above the axios request,
we do not apply these headers if the request is doing some dangerous
things like loopbacks or accessing LAN when it was denied.

Update the origin header check to introduce the setting
corsApplyHeadersRegardlessOfOriginHeader to determine if the
_applyCorsHeaders method should be invoked even if the origin header is
not present in the request.

Remove the `Access-Control-Allow-Origin` assignment from both the axios
.then() and .catch(), as it is set further up the pipeline.

Add a check for allowCorsHeaderOverwrite (a local variable that is
returned by _applyCorsHeaders and is true by default according to it's
declaration), if true this will only overwrite the
Access-Control-Allow-Headers, Access-Control-Allow-Methods,
Access-Control-Max-Age and Access-Control-Expose-Headers if they are
present in the axios response object.

A slight change to the dispatch of the response to the client, just
before we end the response, transform the keys of every header to
lowercase. This however will not write to Date or Connection headers.