Update sphynxServiceRewriteReader.ts
A slight change to the implementation of this, update the initialize()
method to a private method and rename it to _initialize().

Add fallbacks to the JSON and YAML reads for both rewrite rules and
hardcodedResponseRules.

Remove a lot of the for .. of .. loops and merge them all into one for
harcodedResponseRules.