RBXPRR-20: Custom Cors Configurations (#40)
[SKIP CI][SKIP ANALYSIS]

* RBXPRR-20: Add corsWriter.ts

This class boasts the ability to return data for route-specific cors
configurations.

Such as:

Configuration
```yml
- routeTemplate: \/?url\/?
  hostname: *
  allowedHeaders
    - Test-Header
  allowedOrigins:
    - ^(http|https):\/\/(localhost|127(\.\d{0,3}){3})(:\d+)?\/?$
```

~ Request No.1
POST /url/
Origin: https://localhost:5000

~ Request No.2
POST /abc/
Origin: https://test.local/abc/123

Request No.1 will respond with the CORs headers, whereas Request No.2
won't.

The two reasons Request No.1 will return the CORs headers is:
  1. The request path  matches the `routeTemplate` RegExp in the configuration, be that it matches 'url' with an optional leading and trailing slash.
  2. The request header 'Origin' matched one of the `allowedOrigins`
     RegExp, allowing it to actually respond.

* RBXPRR-20: Update environment.ts

Slight change with _getSettingOrDefault<T>(string, T | () => T,
boolean): Update the typeof the default value to also resolve to a
function so we can guess the response. With this change update both
sslCertificateFileName and sslKeyFileName.

Add the following settings:
 1. corsRulesFileName - The fileName of the CORs rules configuration,
    this file can either be json or yaml. It can only have the
    extensions of `.json`, `.yml` and `.yaml`. This will default to
    'cors-rules.yml'.
 2. corsRulesBaseDirectory - The base directory to search for CORs rule
    configuration files, this will default to
    directories.projectDirectoryName.
 3. corsRulesReloadOnRequest - Determines if on each request to get a
    CORs rule configuration, we should reload the already existing
    configurations. This is not advised to be used in production as it
    may be an expensive operation to perform. Defaults to false.
 4. corsApplyHeadersRegardlessOfOrigin - This setting will determine if
    we should still apply the CORs headers even if the rule's
    allowedOrigins does not match the request origin, and the rule's
    allowRequestOriginIfNoAllowedOrigins field is set to false. Defaults
    to false.
 5. corsApplyHeadersRegardlessOfOriginHeader - This setting will
    determine if _applyCorsHeaders is called, regardless of if the
    Origin header is present or not. Defaults to false.

* RBXPRR-20: Update __all.ts

Add a notice for RBXPRR-27 (Support websockets) at the top of the file.

Add a method _isOriginAllowed that iterates a CORs rule's allowedOrigins
and tests each regex against the request origin.

Rewrite _applyCorsHeaders to now use corsWriter, it will do the
following:
  1. Fetch the CORs rule from corsWriter passing the context's `Request`
     object as an argument (used to determine request method, path etc.)
  2. If the rule does not exist we will return a true value, the return
     of this method determines if further down the proxy pipeline if the
     response from the proxy can overwrite these headers (except
     Access-Control-Allow-Origin), this will also return true if the
     CORs rule's allowResponseHeadersOverwrite is true.
  3. We will call on _isOriginAllowed, passing in the CORs rule's
     allowedOrigins and casting them to a RegExp[] (as it's a (string |
     RegExp)[]) and the request origin we captured from an outside
     context. If the result of this method is false, we will then check
     the CORs rule's allowRequestOriginIfNoAllowedOrigins field and if
     it's true we will proceed to Step 4, otherwise we will check if the
     setting corsApplyHeadersRegardlessOfOrigin is enabled, or we will
     just go back to Step 2.
  4. We can now apply the origin header from the client, this may be
     volatile if corsApplyHeadersRegardlessOfOriginHeader is enabled,
     and the request didn't actually have an Origin header, as this
     header in the response will just be `undefined` as a string. We
     then check if the allowedHeaders, allowedMethods, exposedHeaders,
     maxAge and allowCredentials are set on the CORs rule, and then
     apply the corresponding headers on the response. We will also set
     the `Vary` header to `Origin`, regardless of if the origin is
     actually present.
  5. Return the CORs rule's allowResponseHeadersOverwrite field.

Move the origin header check from the top of the invoke method (around
where the publicIp is initialized to directly above the axios request,
we do not apply these headers if the request is doing some dangerous
things like loopbacks or accessing LAN when it was denied.

Update the origin header check to introduce the setting
corsApplyHeadersRegardlessOfOriginHeader to determine if the
_applyCorsHeaders method should be invoked even if the origin header is
not present in the request.

Remove the `Access-Control-Allow-Origin` assignment from both the axios
.then() and .catch(), as it is set further up the pipeline.

Add a check for allowCorsHeaderOverwrite (a local variable that is
returned by _applyCorsHeaders and is true by default according to it's
declaration), if true this will only overwrite the
Access-Control-Allow-Headers, Access-Control-Allow-Methods,
Access-Control-Max-Age and Access-Control-Expose-Headers if they are
present in the axios response object.

A slight change to the dispatch of the response to the client, just
before we end the response, transform the keys of every header to
lowercase. This however will not write to Date or Connection headers.

* Update sphynxServiceRewriteReader.ts

A slight change to the implementation of this, update the initialize()
method to a private method and rename it to _initialize().

Add fallbacks to the JSON and YAML reads for both rewrite rules and
hardcodedResponseRules.

Remove a lot of the for .. of .. loops and merge them all into one for
harcodedResponseRules.

* RBXPRR-20: Update Dockerfile

Add new environment variables, while also including the ones I forgot
previously.

Remove the COPY for package*.json as it's already copied, this was an
artefact from when I would have ran typescript-compile within this
Dockerfile

* RBXPRR-20: Update .gitignore

Add the cors-rules.[yml|yaml|json] files to the ignore file, please read
cors-rules.example.yml for more information on this format

* RBXPRR-20: Update .dockerignore

Allow the cors-rules to be written to the docker images so that custom
rules can be fetched at runtime

* RBXPRR-20: Update __all.ts

Small change to the implementation of _applyCorsHeaders. If the CORs
rule's allowedOrigins includes /\*/ then we can write
'Access-Control-Allow-Origin: *'

* RBXPRR-20: Update __all.ts

Another change here to how this Access-Control-Allow-Origin header is
written, instead we will match /^\*$/, and we will also check in
_isOriginAllowed if the allowedOrigin iterator is equal to this. Within
the corsWriter, if the allowedOrigin '*', we will trasform it to /^\*$/
and remove all the other allowedOrigins

* RBXPRR-20: Update corsWriter.ts

Little change here on the modification of allowedOrigins, if the
allowedOrigins includes a single string of '*', change that to /^\*$/
and remove every other origin, also check if `o` is an instanceof
'RegExp' already when transforming the origins to RegExps

* RBXPRR-20: Add an example cors rules file.

This will include some comments on the information of each key