RBXPRR-30: Update __all.ts
Update _isOriginAllowed to check if the allowedOrigin's string matches
'/^\\*$/' because for some reason you cannot compare RegExps, if so
remove that RegExp and replace it with '*' so that _applyCorsHeaders can
read it.

Update _applyCorsHeaders to check if the CORs rule's allowedOrigins
includes '*' as the _isOriginAllowed would have transformed it. And if
not only apply Access-Control-Allow-Origin if the captured request
origin is not undefined.

Use request.originalUrl instead of request.url.

Move the Sphynx writer stuff down just above the axios request and below
the CORs headers application so that we can have CORs headers with the
harcoded responses.

Note: I could probably shorten this a bit more by just ignoring the '*'
within the allowedOrigins in the corsWriter.