RBXPRR-34 RBXPRR-29: Reverse Proxy middleware
This middleware’s primary function is to resolve client sided proxy
headers like x-forwarded-for eytc. It boasts the ability for client IPv4
and IPv6 validation by checking if the client is LAN (I want to change
this to optional, as the functionality to check if the client is from
rfc1918 etc is forced, while we actually have a setting for
“allowedServerIps…”, which means we can validate individual, known
reverse proxies instead of trusting every address within our LAN),
cloudflare IP translation via CF-Forwarded-IP (there’s a setting with a
list of IPs that act as cloudflare’s servers, I could just put this in
the allowedServerIps, but with this it has the default of cloudflare’s
official IP list), and a list of IPv4 and IPv6 addresses that act as the
“trusted reverse proxies” in case you’re requesting from somewhere other
than the LAN (see LAN comment), translation of the request context IP
based on a dynamic header name (default x-forwarded-for), support for
translation of request host, although I don’t recommend this as you
should just make your reverse proxies translate it via the host header
instead, translation of the client’s scheme (will support it for cf
also), and support for client translation of the socket’s port, although
I’ll have to figure out what to assign on the request, as assigning
request.socket.localPort is not really a viable option.

Co-Authored-by: Leighton Kavanagh <convex@rkaev.net>