RBXPRR-39 RBXPRR-29: WAN Address Application Middleware
This middleware’s primary function is to set the current request’s WAN
address that is used by the loopback check middleware.

The value here is cached at will stay the same throughout the entire
application runtime, so if you happen to change that IP, restart the
application.

I may support reloading of this address in the near future, but for now
this is the best you’ll get!