RBXPRR-40 RBXPRR-29: Hostname Resolution Middleware
This middleware’s primary function is to try get, and validate the
request hostname, as well as transform it if it matches
environment.robloxTestSiteDomainRegex.

Another thing this can do is strip the port out of the hostname, as some
clients, while it being a low chance, will send the request with the
port in the host header.