RBXPRR-41 RBXPRR-29: Deny Local Area Network Access Middleware
This middleware’s primary function is to determine if the address
resolved from the hostnameResolutionMiddlware matches an RFC1918,
RFC4193, RFC3879 or Link-Local address.

This doesn’t check if theoopback middleware.

This middleware will not ex address is loopback though, that is done in
the lecute if environment.hateLocalAreaNetworkAccess is false.