RBXPRR-44 RBXPRR-29: Sphynx Domain Middleware
This middleware’s primary function is to respond wllows it, or to
transform the request url’s servicith a hardcoded response if there is a
rule that ae name.

I want to stray away from the hardcoding being Sphynx-specific, I will
implement a thing that will allow me to specify the hostname like the
corsWriter, and will probably move it.