RBXPRR-45 RBXPRR-29: Send Axios Request Middleware
This middleware’s primary function is to send the
actual proxied request to the downstream.

This has been changed since to just fetch the response as a stream
instead of an arrayBuffer, as there was sizing issues with the response.

This fixes https://mfdlabs.atlassian.net/browse/RBXPRR-26 but not
https://mfdlabs.atlassian.net/browse/RBXPRR-35 just yet.